Refactor CLI code
