+++ packages/uni-watcher/src/cli/reset-cmds/state.ts
@@ -6,7 +6,7 @@ import debug from 'debug';
 import { MoreThan } from 'typeorm';
 import assert from 'assert';
 
-import { getConfig, getResetConfig, resetJobs } from '@vulcanize/util';
+import { getConfig, initClients, resetJobs } from '@vulcanize/util';
 
 import { Database } from '../../database';
 import { Indexer } from '../../indexer';
@@ -27,7 +27,7 @@ export const builder = {
 export const handler = async (argv: any): Promise<void> => {
   const config = await getConfig(argv.configFile);
   await resetJobs(config);
-  const { dbConfig, ethClient, postgraphileClient } = await getResetConfig(config);
+  const { dbConfig, ethClient, postgraphileClient } = await initClients(config);
 
   // Initialize database.
   const db = new Database(dbConfig);
