+++ packages/util/index.ts
@@ -13,3 +13,4 @@ export * from './src/types';
 export * from './src/indexer';
 export * from './src/job-runner';
 export * from './src/ipldHelper';
+export * from './src/cli';
